google-adk/tour-of-agents/pyproject.toml
[project]
name = "tour-of-agents"
version = "0.1.0"
description = "Tour of Google ADK agents using Restate for durability"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hypercorn>=0.17.3",
    "pydantic>=2.10.6",
    "restate_sdk[serde]>=0.13.2",
    "google-adk>=1.18.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
